refactor styles for BookCard
